requirements.txt

pandas==2.2.2
python-telegram-bot==21.6
requests==2.32.3
beautifulsoup4==4.12.3

